+++ basyx.common/basyx.authorization/src/test/java/org/eclipse/digitaltwin/basyx/authorization/AccessTokenProvider.java
@@ -66,6 +66,7 @@ public class AccessTokenProvider {
 
 			CloseableHttpClient httpClient = HttpClients.createDefault();
 
+			@SuppressWarnings("deprecation")
 			CloseableHttpResponse httpResponse = httpClient.execute(httpPost);
 
 			return getAccessTokenFromResponse(httpResponse);
